fixes references

+++ src/main/java/project/transfi/repository/StatusRepository.java
@@ -3,5 +3,5 @@ package project.transfi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import project.transfi.entity.Status;
 
-public interface StatusRepository extends JpaRepository<Status, Integer> {
+public interface StatusRepository extends JpaRepository<Status, Long> {
 }
